refactor: 팀 컨벤션에 맞춰 수정
