+++ src/main/java/com/woowacamp/soolsool/core/liquor/domain/vo/LiquorBrewType.java
@@ -1,8 +1,5 @@
 package com.woowacamp.soolsool.core.liquor.domain.vo;
 
-import static com.woowacamp.soolsool.global.exception.LiquorErrorCode.NOT_LIQUOR_BREW_TYPE_FOUND;
-
-import com.woowacamp.soolsool.global.exception.SoolSoolException;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
@@ -22,11 +19,4 @@ public enum LiquorBrewType {
     private final String type;
     private final String eType;
 
-    public static LiquorBrewType findType(final String name) {
-        try {
-            return valueOf(name);
-        } catch (IllegalArgumentException e) {
-            throw new SoolSoolException(NOT_LIQUOR_BREW_TYPE_FOUND);
-        }
-    }
 }
