+++ src/main/java/com/woowacamp/soolsool/core/liquor/domain/vo/LiquorRegionType.java
@@ -1,8 +1,5 @@
 package com.woowacamp.soolsool.core.liquor.domain.vo;
 
-import static com.woowacamp.soolsool.global.exception.LiquorErrorCode.NOT_LIQUOR_REGION_TYPE_FOUND;
-
-import com.woowacamp.soolsool.global.exception.SoolSoolException;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
@@ -24,11 +21,4 @@ public enum LiquorRegionType {
     private final String name;
     private final String eName;
 
-    public static LiquorRegionType findType(final String name) {
-        try {
-            return valueOf(name);
-        } catch (IllegalArgumentException e) {
-            throw new SoolSoolException(NOT_LIQUOR_REGION_TYPE_FOUND);
-        }
-    }
 }
