+++ src/main/java/com/woowacamp/soolsool/core/liquor/domain/vo/LiquorStatusType.java
@@ -1,8 +1,5 @@
 package com.woowacamp.soolsool.core.liquor.domain.vo;
 
-import static com.woowacamp.soolsool.global.exception.LiquorErrorCode.NOT_LIQUOR_STATUS_TYPE_FOUND;
-
-import com.woowacamp.soolsool.global.exception.SoolSoolException;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
@@ -16,12 +13,4 @@ public enum LiquorStatusType {
 
     private final String status;
     private final String eStatus;
-
-    public static LiquorStatusType findType(final String name) {
-        try {
-            return valueOf(name);
-        } catch (IllegalArgumentException e) {
-            throw new SoolSoolException(NOT_LIQUOR_STATUS_TYPE_FOUND);
-        }
-    }
 }
