+++ src/main/java/com/woowacamp/soolsool/core/liquor/dto/LiquorSaveRequest.java
@@ -1,0 +1,38 @@
+package com.woowacamp.soolsool.core.liquor.dto;
+
+import com.woowacamp.soolsool.core.liquor.domain.Liquor;
+import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorRegion;
+import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorStatus;
+import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorType;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class LiquorSaveRequest {
+
+    private final String typeName;
+    private final String regionName;
+    private final String statusName;
+    private final String name;
+    private final String price;
+    private final String brand;
+    private final String imageUrl;
+    private final Integer stock;
+    private final Double alcohol;
+    private final Integer volume;
+
+    public Liquor toEntity(
+        final LiquorType liquorType,
+        final LiquorRegion liquorRegion,
+        final LiquorStatus liquorStatus
+    ) {
+
+        return new Liquor(
+            liquorType, liquorRegion,
+            liquorStatus, name,
+            price, brand, imageUrl,
+            stock, alcohol, volume
+        );
+    }
+}
