+++ src/main/java/com/woowacamp/soolsool/core/liquor/repository/LiquorRegionRepository.java
@@ -2,12 +2,11 @@ package com.woowacamp.soolsool.core.liquor.repository;
 
 import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorRegion;
 import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorRegionType;
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface LiquorRegionRepository extends JpaRepository<LiquorRegion, Long> {
 
-    Optional<LiquorRegion> findByType(final LiquorRegionType type);
+    LiquorRegion findByType(final LiquorRegionType type);
 }
