+++ src/main/java/com/woowacamp/soolsool/core/liquor/repository/LiquorStatusRepository.java
@@ -2,12 +2,11 @@ package com.woowacamp.soolsool.core.liquor.repository;
 
 import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorStatus;
 import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorStatusType;
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface LiquorStatusRepository extends JpaRepository<LiquorStatus, Long> {
 
-    Optional<LiquorStatus> findByType(final LiquorStatusType type);
+    LiquorStatus findByType(final LiquorStatusType type);
 }
