+++ src/main/java/com/woowacamp/soolsool/core/liquor/repository/LiquorTypeRepository.java
@@ -2,12 +2,11 @@ package com.woowacamp.soolsool.core.liquor.repository;
 
 import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorBrewType;
 import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorType;
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface LiquorTypeRepository extends JpaRepository<LiquorType, Long> {
 
-    Optional<LiquorType> findByType(final LiquorBrewType type);
+    LiquorType findByType(final LiquorBrewType type);
 }
