+++ src/test/java/com/woowacamp/soolsool/acceptance/LiquorAcceptanceTest.java
@@ -4,13 +4,15 @@ import static com.woowacamp.soolsool.global.common.LiquorResultCode.LIQUOR_CREAT
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.http.HttpStatus.CREATED;
 
-import com.woowacamp.soolsool.acceptance.helper.RestAssuredHelper.Liquor;
-import com.woowacamp.soolsool.core.liquor.dto.SaveLiquorRequest;
+import com.woowacamp.soolsool.core.liquor.dto.LiquorSaveRequest;
 import com.woowacamp.soolsool.global.common.ApiResponse;
+import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
 
 @DisplayName("술 상품 관련 기능")
 class LiquorAcceptanceTest extends AcceptanceTest {
@@ -20,7 +22,7 @@ class LiquorAcceptanceTest extends AcceptanceTest {
     void saveLiquor() {
         // given
         String accessToken = "1234";
-        SaveLiquorRequest saveLiquorRequest = new SaveLiquorRequest(
+        LiquorSaveRequest liquorSaveRequest = new LiquorSaveRequest(
             "SOJU",
             "GYEONGGI_DO",
             "ON_SALE",
@@ -32,10 +34,15 @@ class LiquorAcceptanceTest extends AcceptanceTest {
             300);
 
         // when
-        ExtractableResponse<Response> response = Liquor.saveLiquor(
-            accessToken,
-            saveLiquorRequest
-        );
+        ExtractableResponse<Response> response = RestAssured
+            .given().log().all()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .accept(MediaType.APPLICATION_JSON_VALUE)
+            .header(HttpHeaders.AUTHORIZATION, accessToken)
+            .body(liquorSaveRequest)
+            .when().post("/liquors")
+            .then().log().all()
+            .extract();
 
         // then
         assertThat(response.statusCode()).isEqualTo(CREATED.value());
